requirements.txt
flask
flask-cors
python-dotenv
elevenlabs
gunicorn
werkzeug